fix: loading from cache in some cases #290
