+++ frontend/lib/components/memberPartCard.dart
@@ -1,8 +1,4 @@
-import 'dart:html';
-
 import 'package:flutter/material.dart';
-import 'package:frontend/services/authService.dart';
-import 'package:provider/provider.dart'; 
 
 final Map<String, String> roles = {
   "MEMBER": "Member",
