+++ frontend/lib/models/member.dart
@@ -1,7 +1,6 @@
 import 'dart:convert';
 
 import 'package:frontend/models/contact.dart';
-import 'package:frontend/models/team.dart';
 
 class Member {
   final String id;
