+++ frontend/lib/routes/meeting/MeetingPage.dart
@@ -6,25 +6,14 @@ import 'package:frontend/routes/meeting/MeetingsNotifier.dart';
 import 'package:frontend/services/meetingService.dart';
 import 'package:provider/provider.dart';
 
-class MeetingPage extends StatelessWidget {
+class MeetingPage extends StatefulWidget {
   const MeetingPage({Key? key}) : super(key: key);
 
   @override
-  Widget build(BuildContext context) {
-    return Container(
-      child: MeetingList(),
-    );
-  }
+  _MeetingPageState createState() => _MeetingPageState();
 }
 
-class MeetingList extends StatefulWidget {
-  const MeetingList({Key? key}) : super(key: key);
-
-  @override
-  _MeetingListState createState() => _MeetingListState();
-}
-
-class _MeetingListState extends State<MeetingList>
+class _MeetingPageState extends State<MeetingPage>
     with AutomaticKeepAliveClientMixin, SingleTickerProviderStateMixin {
   final MeetingService _service = MeetingService();
   late final Future<List<Meeting>> _meetings;
@@ -38,7 +27,6 @@ class _MeetingListState extends State<MeetingList>
   }
 
   @override
-  // TODO: implement wantKeepAlive
   bool get wantKeepAlive => true;
 
   @override
