+++ frontend/lib/routes/member/EditContact.dart
@@ -195,7 +195,7 @@ class _MyFormState extends State<EditContact> {
                             padding: const EdgeInsets.all(8.0),
                             child: ElevatedButton(
                               style: ElevatedButton.styleFrom(
-                                primary: Theme.of(context).colorScheme.secondary,
+                                backgroundColor: Theme.of(context).colorScheme.secondary,
                                 padding: EdgeInsets.symmetric(horizontal: 50),
                                 elevation: 2,
                                 shape: RoundedRectangleBorder(
